extract arthas_api_common.py
